improvements

+++ Ex05.01/Dogs.c
@@ -50,7 +50,7 @@ int NameComparator(Dog *dog1, Dog *dog2)
     return strcmp(dog1->name, dog2->name);
 }
 
-int SizeComparator(Dog *dog1, Dog *dog2)
+double SizeComparator(Dog *dog1, Dog *dog2)
 {
     return dog1->kg - dog2->kg;
 }
@@ -60,7 +60,7 @@ void PrintArray(Dog *dogs, size_t size)
     printf_s("{");
     for (size_t i = 0; i < size; ++i)
     {
-        printf_s(" [%s, %d], ", dogs[i].name, dogs[i].kg);
+        printf_s(" [%s, %lf], ", dogs[i].name, dogs[i].kg);
     }
     printf_s("}\n");
 }
